+++ CHANGELOG.rst
@@ -31,7 +31,8 @@ Added
     - ``checksampleaccess`` command debug logging (#2261)
     - ``syncstudytables`` ``check`` arg (#1794)
     - ``SheetContextAjaxView`` assay details and comments (#1827)
-    - ``TableDetailModal`` for displaying sheet table details (#1827)
+    - ``TableDetailModal`` for displaying study and assay details (#1827, #2265)
+    - ``SheetContextAjaxView`` additional study details (#2265)
 - **Taskflowbackend**
     - Project viewer role support (#2200)
     - ``user`` arg for ``TaskflowAPI.submit()`` (#2157)